shell: input: refactor plugin and remove multiple eventlog watches
Problem: The shell input plugin is difficult to follow and maintain
because it contains three different components:

 1. An "input service" which handles standard input data from an
    external source and shuttles that data into the guest.input eventlog.

 2. A file input handler that reads a file and puts that data into
    the input eventlog.

 3. A set of eventlog watchers (one per task) that watch the input
    eventlog, process data events, and write data to matching tasks.

Move the input plugin to a src/shell/input subdir and split these
3 components into separate input handling plugins:

 1. input/service.c: input service

 2. input/file.c: input file reader

 3. input/kvs.c: guest.input KVS eventlog watcher

While refactoring, the KVS watcher was changed to watch the guest.input
eventlog once per shell instead of once per task. Otherwise, there should
be no overall functionality change with this refactor.
