+++ src/shell/builtins.c
@@ -35,7 +35,9 @@ extern struct shell_builtin builtin_files;
 extern struct shell_builtin builtin_stage_in;
 extern struct shell_builtin builtin_log_eventlog;
 extern struct shell_builtin builtin_pmi;
-extern struct shell_builtin builtin_input;
+extern struct shell_builtin builtin_input_service;
+extern struct shell_builtin builtin_file_input;
+extern struct shell_builtin builtin_kvs_input;
 extern struct shell_builtin builtin_output;
 extern struct shell_builtin builtin_kill;
 extern struct shell_builtin builtin_signals;
@@ -61,7 +63,9 @@ static struct shell_builtin * builtins [] = {
     &builtin_stage_in,
     &builtin_log_eventlog,
     &builtin_pmi,
-    &builtin_input,
+    &builtin_input_service,
+    &builtin_file_input,
+    &builtin_kvs_input,
     &builtin_output,
     &builtin_kill,
     &builtin_signals,
